move role to give logic to payment callback
